CrashServer/requirements.txt
redis
pika
mysql-connector-python
janus
websockets
pymongo
msgpack
pyyaml